Move non-shadcn components out of components/ui into components/common.
Reserves components/ui/ for shadcn-managed primitives. formatted-date,
graph, page-template, and status-indicator are project-local and don't
belong to the shadcn registry.
