+++ app/(app)/admin/devices/client-page.tsx
@@ -42,7 +42,7 @@ import {
 	DropdownMenuSeparator,
 	DropdownMenuTrigger,
 } from "@/components/ui/dropdown-menu";
-import { PageTemplate } from "@/components/ui/page-template";
+import { PageTemplate } from "@/components/common/page-template";
 import {
 	Select,
 	SelectContent,
