+++ app/(app)/catalog/page.tsx
@@ -1,4 +1,4 @@
-import { PageTemplate } from "@/components/ui/page-template";
+import { PageTemplate } from "@/components/common/page-template";
 import {
 	fetchCatalog,
 	fetchTrmnlRecipes,
