+++ app/(app)/system-logs/page.tsx
@@ -4,7 +4,7 @@ import { Suspense } from "react";
 import { fetchSystemLogs } from "@/app/actions/system";
 import SystemLogsViewerSkeleton from "@/components/system-logs/system-logs-viewer-skeleton";
 import { Button } from "@/components/ui/button";
-import { PageTemplate } from "@/components/ui/page-template";
+import { PageTemplate } from "@/components/common/page-template";
 import { getDbStatus } from "@/lib/database/utils";
 import { SystemLogsClientPage } from "./client-page";
