+++ app/(app)/tools/page.tsx
@@ -2,7 +2,7 @@ import Link from "next/link";
 import { Suspense } from "react";
 import tools from "@/app/(app)/tools/tools.json";
 import { Badge } from "@/components/ui/badge";
-import { PageTemplate } from "@/components/ui/page-template";
+import { PageTemplate } from "@/components/common/page-template";
 
 // Tool configuration type
 type ToolConfig = {
