+++ components/common/graph.tsx
