+++ components/common/status-indicator.tsx
