+++ components/mixup/mixup-list.tsx
@@ -3,7 +3,7 @@
 import { Edit, LayoutGrid, Trash2 } from "lucide-react";
 import { AspectRatio } from "@/components/ui/aspect-ratio";
 import { Button } from "@/components/ui/button";
-import { FormattedDate } from "@/components/ui/formatted-date";
+import { FormattedDate } from "@/components/common/formatted-date";
 import { getLayoutById } from "@/lib/mixup/constants";
 import {
 	DEFAULT_IMAGE_HEIGHT,
